Stop posting duplicate tweets
Currently this bot does not know if criteria has changed or not. So it
always posts the age criteria to twitter, and lets twitter reject
duplicates.

But twitter allows a duplicate tweet after ~12 hours. So every 12 hours
the bot repeats itself.

For now lets copy the regex the SMS notifications use to prevent
duplicates being sent.