Stop storing should i send regex as secret
I wanted a way of quickly stopping duplicate SMSes being sent to me, so
in 17fbb8 I moved this regex to a secret so I could update it easily.

But lets face it, I'm not going to remember the regex syntax off the top
of my head. So instead I now have a 'tweeting enabled' feature flag I
can use, which is much easier to update from memory.