Include booking link in Tweet
I want it to be easier for people who see a retweet to know how to book,
so I am including a link in the tweet.

I previously skipped the link because I was worried twitter would flag
it as disinformation or spam, but lets try it and find out.